# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GhostNet VPN Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /beacons:
    get:
      operationId: listBeaconAlerts
      tags: [beacons]
      summary: List all beacon intrusion alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, dismissed, all]
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeaconAlertList"
  /beacons/{id}/dismiss:
    post:
      operationId: dismissBeaconAlert
      tags: [beacons]
      summary: Dismiss a beacon alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeaconAlert"
  /beacons/trigger:
    post:
      operationId: triggerBeacon
      tags: [beacons]
      summary: Manually trigger a beacon test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerBeaconBody"
      responses:
        "201":
          description: Beacon triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeaconAlert"
...
  /firewall/rules:
    get:
      operationId: listFirewallRules
      tags: [firewall]
      summary: List all firewall rules
      responses:
        "200":
          description: List of firewall rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirewallRuleList"
    post:
      operationId: createFirewallRule
      tags: [firewall]
      summary: Create a new firewall rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFirewallRuleBody"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirewallRule"
  /firewall/rules/{id}:
    put:
      operationId: updateFirewallRule
      tags: [firewall]
      summary: Update a firewall rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFirewallRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirewallRule"
    delete:
      operationId: deleteFirewallRule
      tags: [firewall]
      summary: Delete a firewall rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /firewall/toggle:
    post:
      operationId: toggleFirewall
      tags: [firewall]
      summary: Enable or disable the firewall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFirewallBody"
      responses:
        "200":
          description: Firewall toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirewallStatus"
  /firewall/blocked-ips:
    get:
      operationId: listBlockedIps
      tags: [firewall]
      summary: List currently blocked IP addresses
      responses:
        "200":
          description: Blocked IPs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedIpList"
    post:
      operationId: blockIp
      tags: [firewall]
      summary: Manually block an IP address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockIpBody"
      responses:
        "201":
          description: IP blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedIp"
  /firewall/blocked-ips/{id}/unblock:
    post:
      operationId: unblockIp
      tags: [firewall]
      summary: Unblock an IP address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: IP unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedIp"
...
components:
  schemas:
...
    BeaconAlert:
      type: object
      properties:
        id:
          type: integer
        nodeId:
          type: integer
        nodeName:
          type: string
        nodeLayer:
          type: string
        attackerIp:
          type: string
        attackerFingerprint:
          type: string
        probeType:
          type: string
          enum: [ping, port_scan, traceroute, packet_sniff, tunnel_probe]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, dismissed]
        silkWebTrapped:
          type: boolean
        rawData:
          type: string
        detectedAt:
          type: string
          format: date-time
      required: [id, nodeId, nodeName, nodeLayer, attackerIp, attackerFingerprint, probeType, severity, status, silkWebTrapped, detectedAt]
    BeaconAlertList:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/BeaconAlert"
        total:
          type: integer
        activeCount:
          type: integer
      required: [alerts, total, activeCount]
    TriggerBeaconBody:
      type: object
      properties:
        nodeId:
          type: integer
        simulatedIp:
          type: string
        probeType:
          type: string
          enum: [ping, port_scan, traceroute, packet_sniff, tunnel_probe]
      required: [nodeId, probeType]
...
    FirewallRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        direction:
          type: string
          enum: [inbound, outbound, both]
        action:
          type: string
          enum: [allow, deny, drop, reject, masquerade, log]
        protocol:
          type: string
          enum: [tcp, udp, icmp, any]
        sourceIp:
          type: string
        sourcePort:
          type: string
        destIp:
          type: string
        destPort:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
        hitCount:
          type: integer
        description:
          type: string
        isIspMasquerade:
          type: boolean
          description: Whether this rule helps mask the ISP and local host IP
        createdAt:
          type: string
          format: date-time
      required: [id, name, direction, action, protocol, priority, enabled, hitCount, isIspMasquerade, createdAt]
    FirewallRuleList:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/FirewallRule"
        total:
          type: integer
        enabledCount:
          type: integer
      required: [rules, total, enabledCount]
    CreateFirewallRuleBody:
      type: object
      properties:
        name:
          type: string
        direction:
          type: string
          enum: [inbound, outbound, both]
        action:
          type: string
          enum: [allow, deny, drop, reject, masquerade, log]
        protocol:
          type: string
          enum: [tcp, udp, icmp, any]
        sourceIp:
          type: string
        sourcePort:
          type: string
        destIp:
          type: string
        destPort:
          type: string
        priority:
          type: integer
        description:
          type: string
        isIspMasquerade:
          type: boolean
      required: [name, direction, action, protocol]
    UpdateFirewallRuleBody:
      type: object
      properties:
        name:
          type: string
        direction:
          type: string
          enum: [inbound, outbound, both]
        action:
          type: string
          enum: [allow, deny, drop, reject, masquerade, log]
        protocol:
          type: string
          enum: [tcp, udp, icmp, any]
        sourceIp:
          type: string
        sourcePort:
          type: string
        destIp:
          type: string
        destPort:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
        description:
          type: string
        isIspMasquerade:
          type: boolean
    FirewallStatus:
      type: object
      properties:
        enabled:
          type: boolean
        mode:
          type: string
          enum: [stealth, strict, standard, learning]
        totalRules:
          type: integer
        enabledRules:
          type: integer
        blockedIps:
          type: integer
        packetsBlocked:
          type: integer
        packetsAllowed:
          type: integer
        ispMasqueradeActive:
          type: boolean
        localhostHidden:
          type: boolean
        dnsMasked:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
      required: [enabled, mode, totalRules, enabledRules, blockedIps, packetsBlocked, packetsAllowed, ispMasqueradeActive, localhostHidden, dnsMasked, lastUpdated]
    ToggleFirewallBody:
      type: object
      properties:
        enabled:
          type: boolean
        mode:
          type: string
          enum: [stealth, strict, standard, learning]
      required: [enabled]
    BlockedIp:
      type: object
      properties:
        id:
          type: integer
        ip:
          type: string
        reason:
          type: string
        autoBlocked:
          type: boolean
        blockedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        hitCount:
          type: integer
      required: [id, ip, reason, autoBlocked, blockedAt, hitCount]
    BlockedIpList:
      type: object
      properties:
        blockedIps:
          type: array
          items:
            $ref: "#/components/schemas/BlockedIp"
        total:
          type: integer
      required: [blockedIps, total]
    BlockIpBody:
      type: object
      properties:
        ip:
          type: string
        reason:
          type: string
        expiresInMinutes:
          type: integer
      required: [ip, reason]
...