# lib/api-zod/src/generated/types/firewallStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostNet VPN Management API
 * OpenAPI spec version: 0.1.0
 */
import type { FirewallStatusMode } from "./firewallStatusMode";

export interface FirewallStatus {
  enabled: boolean;
  mode: FirewallStatusMode;
  totalRules: number;
  enabledRules: number;
  blockedIps: number;
  packetsBlocked: number;
  packetsAllowed: number;
  ispMasqueradeActive: boolean;
  localhostHidden: boolean;
  dnsMasked: boolean;
  lastUpdated: Date;
}