# lib/api-zod/src/generated/types/toggleFirewallBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostNet VPN Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ToggleFirewallBodyMode } from "./toggleFirewallBodyMode";

export interface ToggleFirewallBody {
  enabled: boolean;
  mode?: ToggleFirewallBodyMode;
}