# lib/api-zod/src/generated/types/updateFirewallRuleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostNet VPN Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateFirewallRuleBodyAction } from "./updateFirewallRuleBodyAction";
import type { UpdateFirewallRuleBodyDirection } from "./updateFirewallRuleBodyDirection";
import type { UpdateFirewallRuleBodyProtocol } from "./updateFirewallRuleBodyProtocol";

export interface UpdateFirewallRuleBody {
  name?: string;
  direction?: UpdateFirewallRuleBodyDirection;
  action?: UpdateFirewallRuleBodyAction;
  protocol?: UpdateFirewallRuleBodyProtocol;
  sourceIp?: string;
  sourcePort?: string;
  destIp?: string;
  destPort?: string;
  priority?: number;
  enabled?: boolean;
  description?: string;
  isIspMasquerade?: boolean;
}